pondering: dcnn pondering experiment
tree search is heavily influenced by initial priors so we can't just reuse
the tree and inject dcnn priors later on. but if we can guess opponent's
next move then we can do some useful work at pondering time: evaluate dcnn
and run tree search ahead of time. later on when play command arrives we
just promote the node and reuse the tree if it matches our guess. if it
doesn't match we discard search results and start from scratch.

so after genmove is over, reset the tree, get dcnn priors for our move and
likely opponent moves (guess from dcnn priors and genmove search results)
and start pondering as usual.

this makes sense when using cpu for dcnn evaluation which is relatively
slow, with a gpu you could evaluate dcnn for every node even during tree
search so no need to go through such restrictions.

number of guesses can be tweaked by 2 uct params:
    dcnn_pondering_prior:  number of guesses from prior best moves (default: 5)
    dcnn_pondering_mcts :  number of guesses from genmove search   (default: 3)

for slow games it makes sense to increase this: we spend more time before
actual search starts but there's more chance we guess right, so that
pondering will be useful. if we guess wrong search results will be
discarded and pondering will not be useful for this move. for fast games
try decreasing it.